Fix warts, such as broken recursion, weird promise chaining, etc.
