+++ lib/common.js
@@ -17,40 +17,33 @@ import Path from 'path';
 import Thing from 'core-util-is';
 
 
-const ENV = {
-    development: /^dev/i,
-    test       : /^test/i,
-    staging    : /^stag/i,
-    production : /^prod/i
-};
+export default {
 
-export default class Common {
+    env: {
+        development: /^dev/i,
+        test       : /^test/i,
+        staging    : /^stag/i,
+        production : /^prod/i
+    },
 
-    static get env() {
-        return ENV;
-    }
-
-    static isAbsolute(path) {
+    isAbsolute(path) {
         if (Thing.isString(path)) {
             path = Path.normalize(path);
             return path === Path.resolve(path);
         }
         return false;
-    }
+    },
 
-    static merge(src, dest) {
+    merge(src, dest) {
         // NOTE: Do not merge arrays and only merge objects into objects.
         if (!Thing.isArray(src) && Thing.isObject(src) && Thing.isObject(dest)) {
-
-            Object.getOwnPropertyNames(src).forEach(function(prop) {
-                var descriptor;
-                descriptor = Object.getOwnPropertyDescriptor(src, prop);
-                descriptor.value = Common.merge(descriptor.value, dest[prop]);
+            for (let prop of Object.getOwnPropertyNames(src)) {
+                let descriptor = Object.getOwnPropertyDescriptor(src, prop);
+                descriptor.value = this.merge(descriptor.value, dest[prop]);
                 Object.defineProperty(dest, prop, descriptor);
-            });
+            }
 
             return dest;
-
         }
 
         return src;
