+++ lib/factory.js
@@ -15,17 +15,45 @@
  \*───────────────────────────────────────────────────────────────────────────*/
 import shush from 'shush'
 import Path from 'path';
+import debuglog from 'debuglog';
 import Config from './config';
 import Common from './common';
 import Handlers from './handlers';
+import Provider from './provider';
 
+const debug = debuglog('confit');
 
 export default class Factory {
 
-    constructor({ basedir, protocols =  {} }) {
+    constructor({ basedir, protocols =  {}, defaults = 'config.json' }) {
         this.basedir = basedir;
         this.protocols = protocols;
-        this.promise = Promise.resolve({});
+        this.promise = Promise.resolve({})
+            .then(store => Common.merge(Provider.argv(), store))
+            .then(store => Common.merge(Provider.env(), store))
+            .then(store => Common.merge(Provider.convenience(), store))
+            .then(Factory.conditional(store => {
+                let file = Path.join(this.basedir, defaults);
+                return Common.merge(shush(file), store)
+            }))
+            .then(Factory.conditional(store => {
+                let file = Path.join(this.basedir, store.env.env + '.json');
+                return Common.merge(shush(file), store);
+            }));
+    }
+
+    static conditional(fn) {
+        return function (store) {
+            try {
+                return fn(store);
+            } catch (err) {
+                if (err.code && err.code === 'MODULE_NOT_FOUND') {
+                    debug('WARNING:', err.message);
+                    return store;
+                }
+                throw err;
+            }
+        }
     }
 
     addDefault(obj) {
